Restructure codebase to kickstart banner ad layout inclusiveness
