+++ app/src/main/java/com/example/goldearn/activities/MainActivity.java
@@ -1,5 +1,8 @@
-package com.example.goldearn;
+package com.example.goldearn.activities;
 
+import com.example.goldearn.App;
+import com.example.goldearn.BuildConfig;
+import com.example.goldearn.R;
 import com.example.goldearn.utils.misc.Configs;
 import com.google.android.gms.ads.AdListener;
 import com.google.android.gms.ads.AdRequest;
@@ -17,7 +20,7 @@ import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
 
-public class BreakGame extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity {
 
     private static final int START_LEVEL = 1;
     private int mLevel;
@@ -33,7 +36,7 @@ public class BreakGame extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_break_game);
+        setContentView(R.layout.activity_main);
 
         // Create the next level button, which tries to show an interstitial when clicked.
         mNextLevelButton = findViewById(R.id.next_level_button);
@@ -95,27 +98,6 @@ public class BreakGame extends AppCompatActivity {
         if (null != mInterstitialAd) {
             mInterstitialAd.loadAd(new AdRequest.Builder().build());
         }
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_break_game, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
     }
 
     private void showInterstitial() {
