+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">Gold Earn</string>
-    <string name="title_activity_break_game">GoldEarning</string>
+    <string name="title_activity">GoldEarning</string>
     <string name="action_settings">Settings</string>
 
     <string name="interstitial_ad_sample">Gold Earnings Admob </string>
